+++ testutil/btcstaking-helper/keeper.go
@@ -677,3 +677,178 @@ func (h *Helper) AddFinalityProvider(fp *types.FinalityProvider) {
 	})
 	h.NoError(err)
 }
+
+func (h *Helper) BuildBTCInclusionProofForSpendingTx(r *rand.Rand, spendingTx *wire.MsgTx, btcHeight uint32) *types.InclusionProof {
+	prevBlockForSpendingTx, _ := datagen.GenRandomBtcdBlock(r, 0, nil)
+	btcHeaderWithProof := datagen.CreateBlockWithTransaction(r, &prevBlockForSpendingTx.Header, spendingTx)
+	btcHeader := btcHeaderWithProof.HeaderBytes
+	btcHeaderInfo := &btclctypes.BTCHeaderInfo{Header: &btcHeader, Height: btcHeight}
+	spendingTxInclusionProof := types.NewInclusionProof(
+		&btcctypes.TransactionKey{Index: 1, Hash: btcHeader.Hash()},
+		btcHeaderWithProof.SpvProof.MerkleNodes,
+	)
+	h.BTCLightClientKeeper.EXPECT().GetHeaderByHash(gomock.Eq(h.Ctx), gomock.Eq(btcHeader.Hash())).Return(btcHeaderInfo, nil).AnyTimes()
+	return spendingTxInclusionProof
+}
+
+func (h *Helper) CreateBtcStakeExpansionWithBtcTipHeight(
+	r *rand.Rand,
+	delSK *btcec.PrivateKey,
+	fpPKs []*btcec.PublicKey,
+	stakingValue int64,
+	stakingTime uint16,
+	prevDel *types.BTCDelegation,
+	lightClientTipHeight uint32,
+) (*wire.MsgTx, *wire.MsgTx, error) {
+	expandMsg := h.createBtcStakeExpandMessage(
+		r,
+		delSK,
+		fpPKs,
+		stakingValue,
+		stakingTime,
+		prevDel,
+	)
+
+	h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: lightClientTipHeight}).MaxTimes(3)
+
+	// Submit the BtcStakeExpand message
+	_, err := h.MsgServer.BtcStakeExpand(h.Ctx, expandMsg)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	spendingTx, err := bbn.NewBTCTxFromBytes(expandMsg.StakingTx)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	fundingTx, err := bbn.NewBTCTxFromBytes(expandMsg.FundingTx)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return spendingTx, fundingTx, nil
+}
+
+// Helper function to create a BtcStakeExpand message for testing
+func (h *Helper) createBtcStakeExpandMessage(
+	r *rand.Rand,
+	delSK *btcec.PrivateKey,
+	fpPKs []*btcec.PublicKey,
+	stakingValue int64,
+	stakingTime uint16,
+	prevDel *types.BTCDelegation,
+) *types.MsgBtcStakeExpand {
+	// Get staking parameters
+	params := h.BTCStakingKeeper.GetParams(h.Ctx)
+
+	// Convert fpPKs to BIP340PubKey format
+	var fpBtcPkList []bbn.BIP340PubKey
+	for _, fpPK := range fpPKs {
+		fpBtcPkList = append(fpBtcPkList, *bbn.NewBIP340PubKeyFromBTCPK(fpPK))
+	}
+
+	// Convert covenant keys
+	var covenantPks []*btcec.PublicKey
+	for _, pk := range params.CovenantPks {
+		covenantPks = append(covenantPks, pk.MustToBTCPK())
+	}
+
+	// Create funding transaction
+	fundingTx := datagen.GenRandomTxWithOutputValue(r, 10000000)
+
+	// Convert previousStakingTxHash to OutPoint
+	prevDelTxHash := prevDel.MustGetStakingTxHash()
+	prevStakingOutPoint := wire.NewOutPoint(&prevDelTxHash, datagen.StakingOutIdx)
+
+	// Convert fundingTxHash to OutPoint
+	fundingTxHash := fundingTx.TxHash()
+	fundingOutPoint := wire.NewOutPoint(&fundingTxHash, 0)
+	outPoints := []*wire.OutPoint{prevStakingOutPoint, fundingOutPoint}
+
+	// Generate staking slashing info using multiple inputs
+	stakingSlashingInfo := datagen.GenBTCStakingSlashingInfoWithInputs(
+		r,
+		h.T(),
+		h.Net,
+		outPoints,
+		delSK,
+		fpPKs,
+		covenantPks,
+		params.CovenantQuorum,
+		stakingTime,
+		stakingValue,
+		params.SlashingPkScript,
+		params.SlashingRate,
+		uint16(params.UnbondingTimeBlocks),
+	)
+
+	slashingPathSpendInfo, err := stakingSlashingInfo.StakingInfo.SlashingPathSpendInfo()
+	h.NoError(err)
+
+	// Sign the slashing tx with delegator key
+	delegatorSig, err := stakingSlashingInfo.SlashingTx.Sign(
+		stakingSlashingInfo.StakingTx,
+		datagen.StakingOutIdx,
+		slashingPathSpendInfo.GetPkScriptPath(),
+		delSK,
+	)
+	h.NoError(err)
+
+	// Serialize the staking tx bytes
+	serializedStakingTx, err := bbn.SerializeBTCTx(stakingSlashingInfo.StakingTx)
+	h.NoError(err)
+
+	stkTxHash := stakingSlashingInfo.StakingTx.TxHash()
+	unbondingValue := uint64(stakingValue) - uint64(params.UnbondingFeeSat)
+
+	// Generate unbonding slashing info
+	unbondingSlashingInfo := datagen.GenBTCUnbondingSlashingInfo(
+		r,
+		h.T(),
+		h.Net,
+		delSK,
+		fpPKs,
+		covenantPks,
+		params.CovenantQuorum,
+		wire.NewOutPoint(&stkTxHash, datagen.StakingOutIdx),
+		uint16(params.UnbondingTimeBlocks),
+		int64(unbondingValue),
+		params.SlashingPkScript,
+		params.SlashingRate,
+		uint16(params.UnbondingTimeBlocks),
+	)
+
+	unbondingTxBytes, err := bbn.SerializeBTCTx(unbondingSlashingInfo.UnbondingTx)
+	h.NoError(err)
+
+	delSlashingTxSig, err := unbondingSlashingInfo.GenDelSlashingTxSig(delSK)
+	h.NoError(err)
+
+	// Create proof of possession
+	stakerAddr := sdk.MustAccAddressFromBech32(prevDel.StakerAddr)
+	pop, err := datagen.NewPoPBTC(stakerAddr, delSK)
+	h.NoError(err)
+
+	fundingTxBz, err := bbn.SerializeBTCTx(fundingTx)
+	h.NoError(err)
+
+	return &types.MsgBtcStakeExpand{
+		StakerAddr:                    prevDel.StakerAddr,
+		Pop:                           pop,
+		BtcPk:                         bbn.NewBIP340PubKeyFromBTCPK(delSK.PubKey()),
+		FpBtcPkList:                   fpBtcPkList,
+		StakingTime:                   uint32(stakingTime),
+		StakingValue:                  stakingValue,
+		StakingTx:                     serializedStakingTx,
+		SlashingTx:                    stakingSlashingInfo.SlashingTx,
+		DelegatorSlashingSig:          delegatorSig,
+		UnbondingValue:                int64(unbondingValue),
+		UnbondingTime:                 params.UnbondingTimeBlocks,
+		UnbondingTx:                   unbondingTxBytes,
+		UnbondingSlashingTx:           unbondingSlashingInfo.SlashingTx,
+		DelegatorUnbondingSlashingSig: delSlashingTxSig,
+		PreviousStakingTxHash:         prevDelTxHash.String(),
+		FundingTx:                     fundingTxBz,
+	}
+}